Web display for the board
Add an HTML/JS layer for easy UI on top of the Scala engine. Use JSON for the query language between the front end and back. I'm using a horribly hack with PHP and a command line shell in order to actually invoke the server; I'm sure there's a better way but will figure it out later.

To load, point the webserver at the html/ directory.